// gravitee-apim-vitest/model/apis.ts
export interface Api {
  id: string;
  contextPath: string;
  description: string;
  endpoint: string;
  name: string;
  version: string;
  state: ApiState;
  visibility: ApiVisibility;
  lifecycle_state: ApiLifecycleState;
  created_at: number;
  updated_at: number;
  owner: string;
  workflow_state: string;
  labels?: string[];
}

export interface Error {
  message: string;
}

export interface ApiDefinition {
  id: string;
  description: string;
  proxy: ApiProxy;
  name: string;
  version: string;
  created_at: number;
  updated_at: number;
  owner: string;
}

export interface ApiProxy {
  context_path: string;
  endpoints: ApiEndpoint[];
}

export interface ApiEndpoint {
  name: string;
  target: string;
  inherit: boolean;
}

export interface PortalApi {
  id: string;
  description: string;
  name: string;
  version: string;
  running: boolean;
  public: boolean;
  draft: boolean;
  created_at: number;
  updated_at: number;
  labels: string[];
  entrypoints: string[];
  categories: string[];
}

export enum ApiErrorCodes {
  API_NOT_FOUND = 'errors.api.notFound',
}

export enum ApiVisibility {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}

export enum ApiLifecycleState {
  CREATED = 'CREATED',
  PUBLISHED = 'PUBLISHED',
  UNPUBLISHED = 'UNPUBLISHED',
  DEPRECATED = 'DEPRECATED',
  ARCHIVED = 'ARCHIVED',
}

export enum ApiWorkflowState {
  IN_REVIEW = 'IN_REVIEW',
  REVIEW_OK = 'REVIEW_OK',
}

export enum ApiState {
  STOPPED = 'STOPPED',
  STARTED = 'STARTED',
}